app/requirements.txt

Click==7.0
Flask==1.0.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.0
Werkzeug==0.15.3
requests==2.22.0
mysql-connector-python==8.0.18
